requirements.txt
pymodbus>=3.6
pyserial>=3.5
rplidar-roboticia>=0.9.5
numpy>=1.26
matplotlib>=3.8
jupyter>=1.0
pi-ina219>=1.4
smbus2>=0.4
